move components out of subdirs and insert heading
there's no paired tests or resources in this model
