+++ front-end/index.html
@@ -24,6 +24,7 @@
     </header>
 
     <main tabindex="0">
+      <div id="heading-container" class="heading__component"></div>
       <div id="signup-container" class="signup__component"></div>
       <div id="profile-container" class="profile__component"></div>
       <div id="bloom-form-container" class="bloom-form__component"></div>
@@ -84,6 +85,12 @@
           >
         </p>
       </section>
+    </template>
+    <!-- Page Heading Component Template -->
+    <template id="heading-template">
+      <header class="page-heading">
+        <h2 data-heading>Page heading</h2>
+      </header>
     </template>
     <!-- Signup Component Template -->
     <template id="signup-template">
