+++ front-end/views/login.mjs
@@ -1,6 +1,6 @@
 import {render, destroy} from "../lib/render.mjs";
 import {state, getLoginContainer} from "../index.mjs";
-import {createLogin, handleLogin} from "../components/login/login.mjs";
+import {createLogin, handleLogin} from "../components/login.mjs";
 
 // Initial load - not logged in
 function loginView() {
